requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
Babel==2.8.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap-form==3.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.1
Pillow==7.0.0
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
python-decouple==3.3
Python==3.7.7
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
Unipath==1.1
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
